requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pyinstaller==6.4.0
pytest==8.0.2
black==24.2.0
pylint==3.1.0=